gcloud/functions/crawl4ai-scraper/requirements.txt
Crawl4AI==0.4.3b1
Flask==3.0.3
nest-asyncio==1.6.0
playwright==1.49.1
beautifulsoup4==4.12.3
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.3
python-dotenv==1.0.1
